This entire class allows us to create a "vec" variable assigned to a class, upon doing so each of the functions allow us to pass in new other vectors Which we can apply mathmatical calculations too.
The classes utilise operator overloading to achieve core mathmatical functions to the vectors(all functions are expressed inline within the code).

NOTE - THis code was taken from the "Game Physics In One Weekend" book , as a trace to continue my own studies, however down the line I plan to tinker with this code for curiosity purposes.